refactor(backend): Reflect previous changes
