+++ backend/alembic/env.py
@@ -8,7 +8,7 @@ from sqlalchemy.ext.asyncio import async_engine_from_config
 
 from alembic import context
 
-from app.config import settings
+from app.core.config import settings
 import app.models
 from app.models.base import Base
